+++ tests/integration/components/t-mail-test.js
@@ -10,15 +10,17 @@ module('Integration | Component | t mail', function(hooks) {
   test('it renders', async function(assert) {
     assert.expect(2);
 
-    await render(hbs`{{t-mail}}`);
+    await render(hbs`
+      <TMail />
+    `);
 
     assert.dom('button').hasText('open mailbox');
 
     // Template block usage:
     await render(hbs`
-      {{#t-mail}}
+      <TMail>
         template block text
-      {{/t-mail}}
+      </TMail>
     `);
 
     assert.dom('button').hasText('open mailbox');
@@ -27,7 +29,9 @@ module('Integration | Component | t mail', function(hooks) {
   test('it creates a mail transformicon with defaults', async function(assert) {
     assert.expect(6);
 
-    await render(hbs`{{t-mail}}`);
+    await render(hbs`
+      <TMail />
+    `);
     percySnapshot(assert);
 
     assert.dom('button').hasAttribute('role', 'button');
@@ -41,7 +45,9 @@ module('Integration | Component | t mail', function(hooks) {
   test('it creates a mail transformicon with `is-open=false`', async function(assert) {
     assert.expect(1);
 
-    await render(hbs`{{t-mail is-open=false}}`);
+    await render(hbs`
+      <TMail @is-open={{false}} />
+    `);
     percySnapshot(assert);
 
     assert.dom('button').hasClass('tcon-transform');
@@ -50,7 +56,11 @@ module('Integration | Component | t mail', function(hooks) {
   test('user can click on the transformicon', async function(assert) {
     assert.expect(2);
 
-    await render(hbs`{{t-mail id="t-mail"}}`);
+    // FIXME: https://github.com/rwjblue/ember-angle-bracket-invocation-polyfill/issues/4#issue-328822657
+    // Currently necessary to use `<TAdd @id="t-add" />` syntax when specifying an `id` but the `aid` prefix shouldn't be necessary -- should be able to say `<TAdd @id="t-add" />`. Fix after this issue is closed.
+    await render(hbs`
+      <TMail @id="t-mail" />
+    `);
 
     assert.dom('#t-mail').hasNoClass('tcon-transform');