+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][unreleased]
 ### Added
   - Use `ember-decorators` style syntax
+  - `@ember/optional-features` addon 
+  - Angle Bracket syntax
 
 ### Updated
   - Tests leverage `qunit-dom` syntax